+++ src/Telegram.Bot/Requests/Stickers/AddStickerToSetRequest.cs
@@ -1,9 +1,7 @@
-﻿using System.Net.Http;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Telegram.Bot.Requests.Abstractions;
 using Telegram.Bot.Types;
-using Telegram.Bot.Types.InputFiles;
 
 // ReSharper disable once CheckNamespace
 namespace Telegram.Bot.Requests
@@ -13,7 +11,7 @@ namespace Telegram.Bot.Requests
     /// can have up to 120 stickers. Returns <c>true</c> on success.
     /// </summary>
     [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
-    public class AddStickerToSetRequest : FileRequestBase<bool>, IUserTargetable
+    public abstract class AddStickerToSetRequest : FileRequestBase<bool>, IUserTargetable
     {
         /// <inheritdoc />
         [JsonProperty(Required = Required.Always)]
@@ -24,20 +22,6 @@ namespace Telegram.Bot.Requests
         /// </summary>
         [JsonProperty(Required = Required.Always)]
         public string Name { get; }
-
-        /// <summary>
-        /// <para>
-        /// <b>PNG</b> image with the sticker, must be up to 512 kilobytes in size, dimensions must
-        /// not exceed 512px, and either width or height must be exactly 512px.
-        /// </para>
-        /// <para>
-        /// Pass a <see cref="Types.InputFiles.InputTelegramFile.FileId"/> as a String to send a file that already
-        /// exists on the Telegram servers, pass an HTTP URL as a String for Telegram to get a file from the Internet,
-        /// or upload a new one using multipart/form-data
-        /// </para>
-        /// </summary>
-        [JsonProperty(Required = Required.Always)]
-        public InputOnlineFile PngSticker { get; }
 
         /// <summary>
         /// One or more emoji corresponding to the sticker
@@ -52,31 +36,20 @@ namespace Telegram.Bot.Requests
         public MaskPosition? MaskPosition { get; set; }
 
         /// <summary>
-        /// Initializes a new request with userId, name, pngSticker and emojis
+        /// 
         /// </summary>
-        /// <param name="userId">User identifier of sticker set owner</param>
+        /// <param name="userId">User identifier</param>
         /// <param name="name">Sticker set name</param>
-        /// <param name="pngSticker">
-        /// <b>PNG</b> image with the sticker, must be up to 512 kilobytes in size, dimensions must
-        /// not exceed 512px, and either width or height must be exactly 512px. Pass a
-        /// <see cref="Types.InputFiles.InputTelegramFile.FileId"/> as a String to send a file
-        /// that already exists on the Telegram servers, pass an HTTP URL as a String for Telegram
-        /// to get a file from the Internet, or upload a new one using multipart/form-data
-        /// </param>
         /// <param name="emojis">One or more emoji corresponding to the sticker</param>
-        public AddStickerToSetRequest(long userId, string name, InputOnlineFile pngSticker, string emojis)
+        protected AddStickerToSetRequest(
+            long userId, 
+            string name, 
+            string emojis)
             : base("addStickerToSet")
         {
             UserId = userId;
             Name = name;
-            PngSticker = pngSticker;
             Emojis = emojis;
         }
-
-        /// <inheritdoc />
-        public override HttpContent? ToHttpContent() =>
-            PngSticker.Content is null
-                ? base.ToHttpContent()
-                : ToMultipartFormDataContent(fileParameterName: "tgs_sticker", inputFile: PngSticker);
     }
 }