+++ src/Telegram.Bot/Requests/Stickers/AddVideoStickerToSetRequest.cs
@@ -1,4 +1,4 @@
-﻿using System.Net.Http;
+using System.Net.Http;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Telegram.Bot.Types.InputFiles;
@@ -7,40 +7,40 @@ using Telegram.Bot.Types.InputFiles;
 namespace Telegram.Bot.Requests
 {
     /// <summary>
-    /// Use this request to add a new animated sticker to a set created by the bot. Static sticker sets
+    /// Use this request to add a new video sticker to a set created by the bot. Static sticker sets
     /// can have up to 120 stickers. Returns <c>true</c> on success.
     /// </summary>
     [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
-    public class AddAnimatedStickerToSetRequest : AddStickerToSetRequest
+    public class AddVideoStickerToSetRequest : AddStickerToSetRequest
     {
         /// <summary>
         /// <b>WEBM</b> video with the sticker, uploaded using multipart/form-data.
-        /// <see href="https://core.telegram.org/stickers#animated-sticker-requirements"/>
+        /// <see href="https://core.telegram.org/stickers#video-sticker-requirements"/>
         /// for technical requirements
         /// </summary>
         [JsonProperty(Required = Required.Always)]
-        public InputFileStream TgsSticker { get; private set; }
+        public InputFileStream WebmSticker { get; private set; }
 
         /// <inheritdoc />
-        /// <param name="tgsSticker">
+        /// <param name="webmSticker">
         /// <b>WEBM</b> video with the sticker, uploaded using multipart/form-data.
-        /// <see href="https://core.telegram.org/stickers#animated-sticker-requirements"/>
+        /// <see href="https://core.telegram.org/stickers#video-sticker-requirements"/>
         /// for technical requirements
         /// </param>
 #pragma warning disable CS1573
-        public AddAnimatedStickerToSetRequest(
+        public AddVideoStickerToSetRequest(
             long userId,
             string name,
-            InputFileStream tgsSticker,
+            InputFileStream webmSticker,
             string emojis)
             : base(userId, name, emojis) =>
-            TgsSticker = tgsSticker;
+            WebmSticker = webmSticker;
 #pragma warning restore CS1573
 
         /// <inheritdoc />
         public override HttpContent? ToHttpContent() =>
-            TgsSticker.Content is not null
-                ? ToMultipartFormDataContent(fileParameterName: "tgs_sticker", inputFile: TgsSticker)
+            WebmSticker.Content is not null
+                ? ToMultipartFormDataContent(fileParameterName: "webm_sticker", inputFile: WebmSticker)
                 : base.ToHttpContent();
     }
 }