+++ src/Telegram.Bot/TelegramBotClientExtensions.cs
@@ -3337,8 +3337,9 @@ namespace Telegram.Bot
 
         /// <summary>
         /// Use this method to upload a .PNG file with a sticker for later use in
-        /// <see cref="CreateNewStickerSetAsync"/>/<see cref="CreateNewAnimatedStickerSetAsync"/> and
-        /// <see cref="AddStickerToSetAsync"/>/<see cref="AddAnimatedStickerToSetAsync"/> methods
+        /// <see cref="CreateNewStaticStickerSetAsync"/>, <see cref="CreateNewAnimatedStickerSetAsync"/>,
+        /// <see cref="CreateNewVideoStickerSetAsync"/>, <see cref="AddStaticStickerToSetAsync"/>,
+        /// <see cref="AddAnimatedStickerToSetAsync"/> and <see cref="AddVideoStickerToSetAsync"/> methods
         /// (can be used multiple times).
         /// </summary>
         /// <param name="botClient">An instance of <see cref="ITelegramBotClient"/></param>
@@ -3387,7 +3388,7 @@ namespace Telegram.Bot
         /// <param name="cancellationToken">
         /// A cancellation token that can be used by other objects or threads to receive notice of cancellation
         /// </param>
-        public static async Task CreateNewStickerSetAsync(
+        public static async Task CreateNewStaticStickerSetAsync(
             this ITelegramBotClient botClient,
             long userId,
             string name,
@@ -3400,7 +3401,7 @@ namespace Telegram.Bot
         ) =>
             await botClient.ThrowIfNull(nameof(botClient))
                 .MakeRequestAsync(
-                    request: new CreateNewStickerSetRequest(userId, name, title, pngSticker, emojis)
+                    request: new CreateNewStaticStickerSetRequest(userId, name, title, pngSticker, emojis)
                     {
                         ContainsMasks = containsMasks,
                         MaskPosition = maskPosition
@@ -3454,9 +3455,60 @@ namespace Telegram.Bot
                     cancellationToken
                 )
                 .ConfigureAwait(false);
+        
+        /// <summary>
+        /// Use this method to create a new animated sticker set owned by a user. The bot will be able to edit
+        /// the sticker set thus created.
+        /// </summary>
+        /// <param name="botClient">An instance of <see cref="ITelegramBotClient"/></param>
+        /// <param name="userId">User identifier of created sticker set owner</param>
+        /// <param name="name">
+        /// Short name of sticker set, to be used in <c>t.me/addstickers/</c> URLs (e.g., <i>animals</i>).
+        /// Can contain only english letters, digits and underscores. Must begin with a letter, can't contain
+        /// consecutive underscores and must end in <i>"_by_&lt;bot username&gt;"</i>. <i>&lt;bot_username&gt;</i>
+        /// is case insensitive. 1-64 characters
+        /// </param>
+        /// <param name="title">Sticker set title, 1-64 characters</param>
+        /// <param name="webmSticker">
+        /// <b>WEBM</b> video with the sticker, uploaded using multipart/form-data. See
+        /// <see href="https://core.telegram.org/stickers#video-sticker-requirements"/>
+        /// for technical requirements
+        /// </param>
+        /// <param name="emojis">One or more emoji corresponding to the sticker</param>
+        /// <param name="containsMasks">Pass <c>true</c>, if a set of mask stickers should be created</param>
+        /// <param name="maskPosition">An object for position where the mask should be placed on faces</param>
+        /// <param name="cancellationToken">
+        /// A cancellation token that can be used by other objects or threads to receive notice of cancellation
+        /// </param>
+        public static async Task CreateNewVideoStickerSetAsync(
+            this ITelegramBotClient botClient,
+            long userId,
+            string name,
+            string title,
+            InputFileStream webmSticker,
+            string emojis,
+            bool? containsMasks = default,
+            MaskPosition? maskPosition = default,
+            CancellationToken cancellationToken = default
+        ) =>
+            await botClient.ThrowIfNull(nameof(botClient))
+                .MakeRequestAsync(
+                    request: new CreateNewVideoStickerSetRequest(
+                        userId: userId, 
+                        name: name, 
+                        title: title, 
+                        webmSticker: webmSticker,
+                        emojis: emojis)
+                    {
+                        ContainsMasks = containsMasks,
+                        MaskPosition = maskPosition
+                    },
+                    cancellationToken
+                )
+                .ConfigureAwait(false);
 
         /// <summary>
-        /// Use this method to add a new sticker to a set created by the bot. Static sticker sets can have up
+        /// Use this method to add a new static sticker to a set created by the bot. Static sticker sets can have up
         /// to 120 stickers.
         /// </summary>
         /// <param name="botClient">An instance of <see cref="ITelegramBotClient"/></param>
@@ -3474,7 +3526,7 @@ namespace Telegram.Bot
         /// <param name="cancellationToken">
         /// A cancellation token that can be used by other objects or threads to receive notice of cancellation
         /// </param>
-        public static async Task AddStickerToSetAsync(
+        public static async Task AddStaticStickerToSetAsync(
             this ITelegramBotClient botClient,
             long userId,
             string name,
@@ -3485,7 +3537,43 @@ namespace Telegram.Bot
         ) =>
             await botClient.ThrowIfNull(nameof(botClient))
                 .MakeRequestAsync(
-                    request: new AddStickerToSetRequest(userId, name, pngSticker, emojis)
+                    request: new AddStaticStickerToSetRequest(userId, name, pngSticker, emojis)
+                    {
+                        MaskPosition = maskPosition
+                    },
+                    cancellationToken
+                )
+                .ConfigureAwait(false);
+        
+        /// <summary>
+        /// Use this method to add a new video sticker to a set created by the bot. Video stickers can be added to
+        /// video sticker sets and only to them. Video sticker sets can have up to 50 stickers
+        /// </summary>
+        /// <param name="botClient">An instance of <see cref="ITelegramBotClient"/></param>
+        /// <param name="userId">User identifier of sticker set owner</param>
+        /// <param name="name">Sticker set name</param>
+        /// <param name="webmSticker">
+        /// <b>TGS</b> animation with the sticker, uploaded using multipart/form-data. See
+        /// <see href="https://core.telegram.org/animated_stickers#technical-requirements"/>
+        /// for technical requirements
+        /// </param>
+        /// <param name="emojis">One or more emoji corresponding to the sticker</param>
+        /// <param name="maskPosition">An object for position where the mask should be placed on faces</param>
+        /// <param name="cancellationToken">
+        /// A cancellation token that can be used by other objects or threads to receive notice of cancellation
+        /// </param>
+        public static async Task AddVideoStickerToSetAsync(
+            this ITelegramBotClient botClient,
+            long userId,
+            string name,
+            InputFileStream webmSticker,
+            string emojis,
+            MaskPosition? maskPosition = default,
+            CancellationToken cancellationToken = default
+        ) =>
+            await botClient.ThrowIfNull(nameof(botClient))
+                .MakeRequestAsync(
+                    request: new AddVideoStickerToSetRequest(userId, name, webmSticker, emojis)
                     {
                         MaskPosition = maskPosition
                     },
@@ -3494,7 +3582,7 @@ namespace Telegram.Bot
                 .ConfigureAwait(false);
 
         /// <summary>
-        /// Use this method to add a new sticker to a set created by the bot. Animated stickers can be added to
+        /// Use this method to add a new animated sticker to a set created by the bot. Animated stickers can be added to
         /// animated sticker sets and only to them. Animated sticker sets can have up to 50 stickers
         /// </summary>
         /// <param name="botClient">An instance of <see cref="ITelegramBotClient"/></param>